requirements.txt
dnspython>=2.3.0
argparse>=1.4.0
colorama>=0.4.6
